refactor: create separate component 'PlayersSearchList'
